replace layout by page
